Make tracking objectstores work w/o initializing
This changes the objectstore to test to see if the config has been
set or not.  if not, then it doesn't try to save/load from disk.

+++ aprsd/utils/objectstore.py
@@ -1,3 +1,4 @@
+import abc
 import logging
 import os
 import pathlib
@@ -9,7 +10,7 @@ from aprsd import config as aprsd_config
 LOG = logging.getLogger("APRSD")
 
 
-class ObjectStoreMixin:
+class ObjectStoreMixin(metaclass=abc.ABCMeta):
     """Class 'MIXIN' intended to save/load object data.
 
     The asumption of how this mixin is used:
@@ -23,6 +24,13 @@ class ObjectStoreMixin:
     When APRSD Starts, it calls load()
     aprsd server -f (flush) will wipe all saved objects.
     """
+    @abc.abstractmethod
+    def is_initialized(self):
+        """Return True if the class has been setup correctly.
+
+        If this returns False, the ObjectStore doesn't save anything.
+
+        """
 
     def __len__(self):
         return len(self.data)
@@ -36,13 +44,16 @@ class ObjectStoreMixin:
             return self.data[id]
 
     def _init_store(self):
-        sl = self._save_location()
-        if not os.path.exists(sl):
-            LOG.warning(f"Save location {sl} doesn't exist")
-            try:
-                os.makedirs(sl)
-            except Exception as ex:
-                LOG.exception(ex)
+        if self.is_initialized():
+            sl = self._save_location()
+            if not os.path.exists(sl):
+                LOG.warning(f"Save location {sl} doesn't exist")
+                try:
+                    os.makedirs(sl)
+                except Exception as ex:
+                    LOG.exception(ex)
+        else:
+            LOG.warning(f"{self.__class__.__name__} is not initialized")
 
     def _save_location(self):
         save_location = self.config.get("aprsd.save_location", None)
@@ -68,38 +79,45 @@ class ObjectStoreMixin:
 
     def save(self):
         """Save any queued to disk?"""
-        if len(self) > 0:
-            LOG.info(f"{self.__class__.__name__}::Saving {len(self)} entries to disk at {self._save_location()}")
-            with open(self._save_filename(), "wb+") as fp:
-                pickle.dump(self._dump(), fp)
-        else:
-            LOG.debug(
-                "{} Nothing to save, flushing old save file '{}'".format(
-                    self.__class__.__name__,
-                    self._save_filename(),
-                ),
-            )
-            self.flush()
+        if self.is_initialized():
+            if len(self) > 0:
+                LOG.info(
+                    f"{self.__class__.__name__}::Saving"
+                    f" {len(self)} entries to disk at"
+                    f"{self._save_location()}",
+                )
+                with open(self._save_filename(), "wb+") as fp:
+                    pickle.dump(self._dump(), fp)
+            else:
+                LOG.debug(
+                    "{} Nothing to save, flushing old save file '{}'".format(
+                        self.__class__.__name__,
+                        self._save_filename(),
+                    ),
+                )
+                self.flush()
 
     def load(self):
-        if os.path.exists(self._save_filename()):
-            try:
-                with open(self._save_filename(), "rb") as fp:
-                    raw = pickle.load(fp)
-                    if raw:
-                        self.data = raw
-                        LOG.debug(
-                            f"{self.__class__.__name__}::Loaded {len(self)} entries from disk.",
-                        )
-                        LOG.debug(f"{self.data}")
-            except (pickle.UnpicklingError, Exception) as ex:
-                LOG.error(f"Failed to UnPickle {self._save_filename()}")
-                LOG.error(ex)
-                self.data = {}
+        if self.is_initialized():
+            if os.path.exists(self._save_filename()):
+                try:
+                    with open(self._save_filename(), "rb") as fp:
+                        raw = pickle.load(fp)
+                        if raw:
+                            self.data = raw
+                            LOG.debug(
+                                f"{self.__class__.__name__}::Loaded {len(self)} entries from disk.",
+                            )
+                            LOG.debug(f"{self.data}")
+                except (pickle.UnpicklingError, Exception) as ex:
+                    LOG.error(f"Failed to UnPickle {self._save_filename()}")
+                    LOG.error(ex)
+                    self.data = {}
 
     def flush(self):
         """Nuke the old pickle file that stored the old results from last aprsd run."""
-        if os.path.exists(self._save_filename()):
-            pathlib.Path(self._save_filename()).unlink()
-        with self.lock:
-            self.data = {}
+        if self.is_initialized():
+            if os.path.exists(self._save_filename()):
+                pathlib.Path(self._save_filename()).unlink()
+            with self.lock:
+                self.data = {}
